Refactor populateCountrySelect function for reusability and add documentation
- created a file js called formBuilder.js and moved the "populateCountrySelect" to it;
- Renamed `populateCountrySelect` to `populateSelectField` to generalize its functionality so that
  it can be used for any category that needs to populate a select field.
- Added `cssSelectorID` parameter to allow the function to populate any select element.
- Updated function to trim option text and handle errors related to file loading.
- Added detailed docstring with usage examples and a note about valid file paths.
- Moved `createOption` helper function to improve modularity.
